Fixes failing test: set state on both Sura and Place
Setting Translation properties doesn't seem to mark a record as
modified. So, this test case previously marked the Place as TDIRTY, so
that it would be saved. However, this test case didn't set the Sura to
TDIRTY.

Now that DC-797 is fixed, and the Sura is no longer marked as TDIRTY by
setting a related Place, the state must be set on both the Sura and the
Place.