DC880 - Adding explicit tableName definition to versionable and searchable to prevent false migration generation
When using model prefixes and the searchable and versionable behaviours, migrations are built that drop and recreate indexes when nothing of the schema has changed. This commit fixes this by explicitly declaring a table name in these behaviours (based on the tablename of the owning class) so that the tablename for generated models is correct. Child classes can still override this value in their extended classes if required.